requirements.txt
Django~=4.0
Pillow==8.0.*
django-debug-toolbar==3.4.0
django-admin-sortable2==2.1.5
